+++ .eslintrc.cjs
@@ -1,25 +1,31 @@
+/*
+ * @Author: ryuhangseong liuhangcheng2002@gmail.com
+ * @Date: 2022-09-25 15:28:08
+ * @LastEditors: ryuhangseong liuhangcheng2002@gmail.com
+ * @LastEditTime: 2022-09-25 15:34:27
+ * @FilePath: \vue3-blog\.eslintrc.cjs
+ * @Description:
+ *
+ * Copyright (c) 2022 by ryuhangseong liuhangcheng2002@gmail.com, All Rights Reserved.
+ */
 /* eslint-env node */
-require('@rushstack/eslint-patch/modern-module-resolution')
+require("@rushstack/eslint-patch/modern-module-resolution");
 
 module.exports = {
   root: true,
-  'extends': [
-    'plugin:vue/vue3-essential',
-    'eslint:recommended',
-    '@vue/eslint-config-typescript',
-    '@vue/eslint-config-prettier'
+  extends: [
+    "plugin:vue/vue3-essential",
+    "eslint:recommended",
+    "@vue/eslint-config-typescript",
+    "@vue/eslint-config-prettier",
   ],
   overrides: [
     {
-      files: [
-        'cypress/e2e/**.{cy,spec}.{js,ts,jsx,tsx}'
-      ],
-      'extends': [
-        'plugin:cypress/recommended'
-      ]
-    }
+      files: ["cypress/e2e/**.{cy,spec}.{js,ts,jsx,tsx}"],
+      extends: ["plugin:cypress/recommended"],
+    },
   ],
   parserOptions: {
-    ecmaVersion: 'latest'
-  }
-}
+    ecmaVersion: "latest",
+  },
+};