+++ src/App.vue
@@ -1,37 +1,34 @@
+<!--
+ * @Author: ryuhangseong liuhangcheng2002@gmail.com
+ * @Date: 2022-09-25 15:50:13
+ * @LastEditors: ryuhangseong liuhangcheng2002@gmail.com
+ * @LastEditTime: 2022-09-25 16:44:10
+ * @FilePath: \vue3-blog\src\App.vue
+ * @Description: 
+ * 
+ * Copyright (c) 2022 by ryuhangseong liuhangcheng2002@gmail.com, All Rights Reserved. 
+-->
 <script setup lang="ts">
-import { RouterLink, RouterView } from 'vue-router'
-import HelloWorld from './components/HelloWorld.vue'
-import { ElButton } from 'element-plus'
-import 'element-plus/es/components/button/style/css'
-
+import { RouterLink, RouterView } from "vue-router";
+import { ElButton } from "element-plus";
+import "element-plus/es/components/button/style/css";
 </script>
-
 <template>
   <header>
-    <img alt="Vue logo" class="logo" src="@/assets/logo.svg" width="125" height="125" />
-
     <div class="wrapper">
-      <HelloWorld msg="You did it!" />
       <el-button>Click here!</el-button>
       <nav>
         <RouterLink to="/">Home</RouterLink>
-        <RouterLink to="/about">About</RouterLink>
       </nav>
     </div>
   </header>
 
   <RouterView />
 </template>
-
 <style scoped>
 header {
   line-height: 1.5;
   max-height: 100vh;
-}
-
-.logo {
-  display: block;
-  margin: 0 auto 2rem;
 }
 
 nav {
@@ -57,32 +54,5 @@ nav a {
 
 nav a:first-of-type {
   border: 0;
-}
-
-@media (min-width: 1024px) {
-  header {
-    display: flex;
-    place-items: center;
-    padding-right: calc(var(--section-gap) / 2);
-  }
-
-  .logo {
-    margin: 0 2rem 0 0;
-  }
-
-  header .wrapper {
-    display: flex;
-    place-items: flex-start;
-    flex-wrap: wrap;
-  }
-
-  nav {
-    text-align: left;
-    margin-left: -1rem;
-    font-size: 1rem;
-
-    padding: 1rem 0;
-    margin-top: 1rem;
-  }
 }
 </style>