+++ src/main.ts
@@ -1,14 +1,22 @@
-import { createApp } from 'vue'
-import { createPinia } from 'pinia'
+/*
+ * @Author: ryuhangseong liuhangcheng2002@gmail.com
+ * @Date: 2022-09-25 15:50:13
+ * @LastEditors: ryuhangseong liuhangcheng2002@gmail.com
+ * @LastEditTime: 2022-09-25 16:25:28
+ * @FilePath: \vue3-blog\src\main.ts
+ * @Description:
+ *
+ * Copyright (c) 2022 by ryuhangseong liuhangcheng2002@gmail.com, All Rights Reserved.
+ */
+import { createApp } from "vue";
+import { createPinia } from "pinia";
 
-import App from './App.vue'
-import router from './router'
+import App from "./App.vue";
+import router from "./router";
 
-import './assets/main.css'
+const app = createApp(App);
 
-const app = createApp(App)
+app.use(createPinia());
+app.use(router);
 
-app.use(createPinia())
-app.use(router)
-
-app.mount('#app')
+app.mount("#app");